reduce to_* function impl duplication
